requirements.txt
# Document Generator Backend Dependencies
# Python 3.12+

Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
Pillow>=12.0.0
PyJWT>=2.8.0
WeasyPrint>=68.0
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.9.0
